+++ student-managment-system/lombok.config
@@ -1,0 +1,2 @@
+lombok.addLombokGeneratedAnnotation = true
+lombok.accessors.chain = true
