+++ student-managment-system/src/main/java/com/paremskis/sms/StudentManagmentSystemApplication.java
@@ -1,0 +1,32 @@
+package com.paremskis.sms;
+
+import com.paremskis.sms.persistence.StudentRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+//@ComponentScan(basePackages = {"com.paremskis.com.paremskis.sms.repository"})
+public class StudentManagmentSystemApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(StudentManagmentSystemApplication.class, args);
+    }
+
+    @Autowired
+    private StudentRepository studentRepository;
+
+//    @Override
+//    public void run(String... args) throws Exception {
+//        Student student1 = new Student("Valters", "Krauze", "krauze@valters.lv");
+//        //	studentRepository.save(student1);
+//        Student student2 = new Student("Janis", "Krauze", "krauze@valters.lv");
+//        //studentRepository.save(student2);
+//        Student student3 = new Student("Zigmars", "jansons", "jansons@zigmars.lv");
+//        //	studentRepository.save(student3);
+//        Student student4 = new Student("Sanjay", "Jadhav", "jadhav@sinjay.lv");
+//        //	studentRepository.save(student4);
+
+//    }
+
+}
