+++ student-managment-system/src/main/java/com/paremskis/sms/config/ApplicationConfig.java
@@ -1,0 +1,7 @@
+package com.paremskis.sms.config;
+
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class ApplicationConfig {
+}
