+++ student-managment-system/src/main/java/com/paremskis/sms/persistence/StudentRepository.java
@@ -1,15 +1,14 @@
-package repository;
+package com.paremskis.sms.persistence;
 
 
 
+import com.paremskis.sms.model.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import entity.Student;
 
 
 @Repository
 public interface StudentRepository extends JpaRepository <Student, Long>{
-	
+
 
 }
