+++ student-managment-system/src/main/java/com/paremskis/sms/rest/controller/StudentController.java
@@ -1,0 +1,23 @@
+package com.paremskis.sms.rest.controller;
+
+import com.paremskis.sms.service.StudentService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class StudentController {
+
+    private final StudentService studentService;
+
+    //handler method to handle list students and return mode view
+    @GetMapping("/students")
+    public String listStudents(Model model) {
+        model.addAttribute("students", studentService.getAllStudents());
+        return "students";
+    }
+
+
+}
