+++ student-managment-system/src/main/java/com/paremskis/sms/service/StudentService.java
@@ -1,8 +1,8 @@
-package service;
+package com.paremskis.sms.service;
 
 import java.util.List;
 
-import entity.Student;
+import com.paremskis.sms.model.Student;
 
 public interface StudentService {
