--- student-managment-system/src/main/java/controller/StudentController.java
@@ -1,27 +1,0 @@
-package controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-import org.springframework.ui.Model;
-import service.StudentService;
-
-@Controller
-public class StudentController {
-	
-	private StudentService studentService;
-
-	public StudentController(StudentService studentService) {
-		super();
-		this.studentService = studentService;
-	}
-	
-	//handler method to handle list students and return mode view
-	@GetMapping("/students")
-	public String listStudents(Model model) {
-		model.addAttribute("students", studentService.getAllStudents());
-		return "students";
-	}
-	
-	
-}
