--- student-managment-system/src/main/java/sms/StudentManagmentSystemApplication.java
@@ -1,36 +1,0 @@
-package sms;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-
-import entity.Student;
-import repository.StudentRepository;
-
-@SpringBootApplication
-@ComponentScan(basePackages = {"com.paremskis.repository"})
-public class StudentManagmentSystemApplication implements CommandLineRunner {
-
-	public static void main(String[] args) {
-		SpringApplication.run(StudentManagmentSystemApplication.class, args);
-	}
-
-	@Autowired
-	private StudentRepository studentRepository ;
-
-	@Override
-	public void run(String... args) throws Exception {
-		Student student1 = new Student("Valters", "Krauze", "krauze@valters.lv");
-	//	studentRepository.save(student1);
-		Student student2 = new Student("Janis", "Krauze", "krauze@valters.lv");
-		//studentRepository.save(student2);
-		Student student3 = new Student("Zigmars", "jansons", "jansons@zigmars.lv");
-	//	studentRepository.save(student3);
-		Student student4 = new Student("Sanjay", "Jadhav", "jadhav@sinjay.lv");
-	//	studentRepository.save(student4);
-
-	}
-
-}
