+++ student-managment-system/src/main/resources/application.properties
@@ -1,6 +1,8 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/sms?useSSL=false&serverTimezone=UTC&useLegacyDateTimeCode=false
-spring.datasource.username=root
-spring.datasource.password=Warszawa25!
+#spring.datasource.url=jdbc:mysql://localhost:3306/sms?useSSL=false&serverTimezone=UTC&useLegacyDateTimeCode=false
+#spring.datasource.username=root
+#spring.datasource.password=Warszawa25!
+
+#TODO check @link{https://www.baeldung.com/spring-boot-h2-database}
 
 #Hibernate
 #spring.jpa.properties.hibernate.dialect = "org.hibernate.dialect.MySQL8InnoDBDialect"
